Extract raw query decoder
